Segregate code into MVC pattern
