+++ index.js
@@ -1,250 +1,92 @@
-let developerElm = document.getElementsByClassName('developer-img')[0];
-developerElm.insertAdjacentHTML('beforeend', `
-<svg width="616" height="517" viewBox="0 0 616 517" fill="none" xmlns="http://www.w3.org/2000/svg">
-<g id="undraw_developer_activity 1" clip-path="url(#clip0_1_36)">
-<path id="Vector" d="M552.512 269.055C552.593 333.028 526.931 394.54 480.921 440.66C478.147 443.459 475.298 446.19 472.374 448.853C467.92 452.936 463.317 456.86 458.563 460.626C414.096 495.928 358.583 515.765 300.981 516.937C299.062 516.979 297.136 517 295.202 517C249.263 517 191.482 493.356 139.047 460.626C134.709 457.919 130.409 455.148 126.148 452.315C124.156 451.003 122.174 449.672 120.202 448.323C99.9782 434.563 80.7819 419.452 62.7546 403.101C50.9514 392.427 39.8832 381.023 29.6203 368.962C-10.6696 321.074 -19.5561 277.925 37.8839 269.055C44.6846 268.008 51.2849 266.889 57.6849 265.698C66.9553 263.982 75.8273 262.128 84.301 260.136C95.7151 257.464 106.439 254.565 116.472 251.441C119.847 250.394 123.145 249.326 126.365 248.238C173.977 232.124 205.844 211.111 227.535 188.312C231.215 184.453 234.606 180.547 237.711 176.595C242.461 170.579 246.776 164.255 250.624 157.669C255.936 148.517 260.398 138.93 263.958 129.021L264.153 128.477C281.412 80.5056 276.161 36.0096 295.202 21.1171C376.593 -42.5492 477.922 48.1176 525.484 155.038C527.881 160.425 530.136 165.845 532.248 171.298C533.197 173.734 534.112 176.174 534.993 178.619C537.31 184.99 539.413 191.378 541.301 197.782C542.025 200.225 542.716 202.665 543.372 205.103C544.075 207.734 544.749 210.358 545.371 212.982C549.999 232.194 552.512 251.141 552.512 269.055Z" fill="#F2F2F2"/>
-<path id="Key_w" d="M366.368 152.135H350.246C349.747 152.135 349.268 152.326 348.915 152.667C348.562 153.007 348.363 153.468 348.363 153.949V169.491C348.363 169.729 348.412 169.964 348.507 170.184C348.602 170.403 348.741 170.603 348.916 170.771C349.09 170.939 349.298 171.072 349.526 171.162C349.755 171.253 349.999 171.299 350.246 171.298H366.368C366.866 171.298 367.344 171.108 367.697 170.77C368.05 170.431 368.249 169.971 368.251 169.491V153.949C368.249 153.469 368.051 153.008 367.698 152.668C367.345 152.328 366.867 152.136 366.368 152.135Z" fill="white"/>
-<path id="Key_r" d="M419.238 152.135H403.116C402.617 152.135 402.138 152.326 401.785 152.667C401.432 153.007 401.233 153.468 401.233 153.949V169.491C401.233 169.729 401.282 169.964 401.377 170.184C401.472 170.403 401.611 170.603 401.786 170.771C401.96 170.939 402.168 171.072 402.396 171.162C402.625 171.253 402.869 171.299 403.116 171.298H419.238C419.736 171.298 420.214 171.108 420.567 170.77C420.92 170.431 421.119 169.971 421.121 169.491V153.949C421.119 153.469 420.921 153.008 420.568 152.668C420.215 152.328 419.737 152.136 419.238 152.135Z" fill="white"/>
-<path id="Vector_2" d="M309.325 125.651H293.203C292.704 125.651 292.225 125.842 291.872 126.182C291.519 126.522 291.32 126.984 291.32 127.465V143.006C291.32 143.244 291.369 143.48 291.464 143.699C291.559 143.919 291.697 144.119 291.872 144.286C292.047 144.454 292.255 144.587 292.483 144.678C292.711 144.768 292.956 144.815 293.203 144.814H309.325C309.823 144.814 310.301 144.624 310.654 144.285C311.007 143.946 311.206 143.486 311.208 143.006V127.465C311.206 126.984 311.007 126.523 310.655 126.183C310.302 125.844 309.824 125.652 309.325 125.651V125.651Z" fill="white"/>
-<path id="Vector_3" d="M313.44 152.135H293.479C292.861 152.135 292.268 152.323 291.831 152.658C291.393 152.992 291.148 153.446 291.147 153.919V169.2C291.148 169.434 291.208 169.665 291.326 169.881C291.443 170.097 291.615 170.294 291.832 170.459C292.048 170.624 292.305 170.754 292.588 170.843C292.87 170.932 293.173 170.978 293.479 170.977H313.44C314.057 170.977 314.649 170.79 315.086 170.457C315.523 170.124 315.77 169.672 315.772 169.2V153.919C315.77 153.446 315.524 152.993 315.087 152.659C314.65 152.325 314.058 152.136 313.44 152.135Z" fill="#E6E6E6"/>
-<path id="Vector_4" d="M309.325 178.619H293.203C292.704 178.619 292.225 178.81 291.872 179.15C291.519 179.491 291.32 179.952 291.32 180.433V195.968C291.319 196.206 291.368 196.442 291.462 196.662C291.557 196.883 291.695 197.083 291.87 197.252C292.045 197.42 292.253 197.554 292.482 197.645C292.71 197.736 292.955 197.782 293.203 197.782H309.325C309.824 197.782 310.303 197.591 310.656 197.25C311.009 196.91 311.207 196.449 311.208 195.968V180.433C311.206 179.952 311.007 179.492 310.655 179.152C310.302 178.812 309.824 178.62 309.325 178.619V178.619Z" fill="#E6E6E6"/>
-<path id="Vector_5" d="M311.179 206.596C311.098 206.178 310.869 205.8 310.531 205.528C310.192 205.255 309.766 205.105 309.325 205.103H293.203C292.704 205.103 292.225 205.294 291.872 205.634C291.519 205.975 291.32 206.436 291.32 206.917V222.452C291.319 222.69 291.368 222.926 291.462 223.146C291.557 223.367 291.695 223.567 291.87 223.736C292.045 223.904 292.253 224.038 292.482 224.129C292.71 224.22 292.955 224.266 293.203 224.266H309.325C309.824 224.266 310.303 224.075 310.656 223.734C311.009 223.394 311.207 222.933 311.208 222.452V206.917C311.209 206.809 311.199 206.702 311.179 206.596V206.596Z" fill="#E6E6E6"/>
-<path id="Vector_6" d="M310.056 231.726C309.824 231.634 309.576 231.586 309.325 231.587H293.203C292.955 231.586 292.71 231.633 292.482 231.724C292.253 231.815 292.045 231.948 291.87 232.117C291.695 232.286 291.557 232.486 291.462 232.706C291.368 232.927 291.319 233.163 291.32 233.401V248.936C291.319 249.174 291.367 249.411 291.461 249.631C291.555 249.852 291.694 250.052 291.869 250.221C292.044 250.39 292.252 250.523 292.481 250.614C292.71 250.705 292.955 250.751 293.203 250.75H309.325C309.824 250.75 310.303 250.559 310.656 250.218C311.009 249.878 311.208 249.417 311.208 248.936V233.401C311.207 233.044 311.097 232.695 310.893 232.397C310.688 232.1 310.397 231.866 310.056 231.726Z" fill="#E6E6E6"/>
-<path id="Vector_7" d="M343.19 125.651H327.069C326.569 125.651 326.091 125.842 325.738 126.182C325.384 126.522 325.186 126.984 325.186 127.465V143.006C325.186 143.244 325.235 143.48 325.33 143.699C325.425 143.919 325.563 144.119 325.738 144.286C325.913 144.454 326.121 144.587 326.349 144.678C326.577 144.768 326.822 144.815 327.069 144.814H343.19C343.689 144.814 344.167 144.624 344.52 144.285C344.873 143.946 345.072 143.486 345.074 143.006V127.465C345.072 126.984 344.873 126.523 344.52 126.183C344.168 125.844 343.689 125.652 343.19 125.651Z" fill="#E6E6E6"/>
-<path id="Key_q" d="M340.295 152.135H324.173C323.674 152.135 323.195 152.326 322.842 152.667C322.489 153.007 322.29 153.468 322.29 153.949V169.491C322.29 169.729 322.339 169.964 322.434 170.184C322.529 170.403 322.668 170.603 322.843 170.771C323.017 170.939 323.225 171.072 323.453 171.162C323.682 171.253 323.926 171.299 324.173 171.298H340.295C340.793 171.298 341.271 171.108 341.624 170.77C341.977 170.431 342.176 169.971 342.178 169.491V153.949C342.176 153.469 341.978 153.008 341.625 152.668C341.272 152.328 340.794 152.136 340.295 152.135Z" fill="white"/>
-<path id="Key_a" d="M343.19 178.619H327.069C326.569 178.619 326.091 178.81 325.737 179.15C325.384 179.491 325.186 179.952 325.186 180.433V195.968C325.185 196.206 325.234 196.442 325.328 196.662C325.423 196.883 325.561 197.083 325.736 197.252C325.911 197.42 326.119 197.554 326.348 197.645C326.576 197.736 326.821 197.782 327.069 197.782H343.19C343.69 197.782 344.169 197.591 344.522 197.25C344.875 196.91 345.073 196.449 345.074 195.968V180.433C345.072 179.952 344.873 179.492 344.52 179.152C344.168 178.812 343.689 178.62 343.19 178.619V178.619Z" fill="white"/>
-<path id="Vector_8" d="M343.916 204.475H327.794C327.4 204.475 327.015 204.595 326.695 204.818C326.375 205.041 326.136 205.356 326.013 205.717C325.946 205.901 325.911 206.094 325.911 206.289V221.824C325.911 222.062 325.959 222.298 326.054 222.519C326.148 222.739 326.287 222.939 326.462 223.108C326.637 223.276 326.844 223.41 327.073 223.501C327.302 223.592 327.547 223.639 327.794 223.638H343.916C344.415 223.638 344.894 223.447 345.247 223.107C345.6 222.766 345.799 222.305 345.799 221.824V206.289C345.798 205.809 345.599 205.348 345.246 205.008C344.893 204.668 344.415 204.476 343.916 204.475V204.475Z" fill="#E6E6E6"/>
-<path id="Vector_9" d="M343.19 231.587H327.069C326.821 231.586 326.576 231.633 326.348 231.724C326.119 231.815 325.911 231.948 325.736 232.117C325.561 232.286 325.423 232.486 325.328 232.706C325.234 232.927 325.185 233.163 325.186 233.401V248.936C325.185 249.174 325.234 249.41 325.328 249.63C325.423 249.851 325.561 250.051 325.736 250.22C325.911 250.388 326.119 250.522 326.348 250.613C326.576 250.704 326.821 250.751 327.069 250.75H343.19C343.69 250.75 344.169 250.559 344.522 250.218C344.875 249.878 345.073 249.417 345.074 248.936V233.401C345.073 232.92 344.875 232.459 344.522 232.118C344.169 231.778 343.69 231.587 343.19 231.587V231.587Z" fill="#E6E6E6"/>
-<path id="Vector_10" d="M377.056 125.651H360.935C360.435 125.651 359.956 125.842 359.603 126.182C359.25 126.522 359.052 126.984 359.052 127.465V143.006C359.052 143.244 359.101 143.48 359.196 143.699C359.29 143.919 359.429 144.119 359.604 144.286C359.779 144.454 359.986 144.587 360.215 144.678C360.443 144.768 360.688 144.815 360.935 144.814H377.056C377.555 144.814 378.032 144.624 378.385 144.285C378.738 143.946 378.938 143.486 378.939 143.006V127.465C378.938 126.984 378.739 126.523 378.386 126.183C378.033 125.844 377.555 125.652 377.056 125.651V125.651Z" fill="#E6E6E6"/>
-<path id="Key_e" d="M392.441 152.135H376.319C375.82 152.135 375.341 152.326 374.988 152.667C374.635 153.007 374.436 153.468 374.436 153.949V169.491C374.436 169.729 374.485 169.964 374.58 170.184C374.675 170.403 374.814 170.603 374.988 170.771C375.163 170.939 375.371 171.072 375.599 171.162C375.827 171.253 376.072 171.299 376.319 171.298H392.441C392.939 171.298 393.417 171.108 393.77 170.77C394.123 170.431 394.322 169.971 394.324 169.491V153.949C394.322 153.469 394.123 153.008 393.771 152.668C393.418 152.328 392.94 152.136 392.441 152.135Z" fill="white"/>
-<path id="Key_s" d="M370.713 178.654H354.591C354.092 178.654 353.613 178.845 353.26 179.186C352.907 179.526 352.709 179.987 352.708 180.468V196.003C352.708 196.241 352.756 196.477 352.851 196.698C352.945 196.918 353.084 197.118 353.259 197.287C353.434 197.455 353.642 197.589 353.87 197.68C354.099 197.771 354.344 197.818 354.591 197.817H370.713C371.213 197.817 371.691 197.626 372.044 197.286C372.398 196.945 372.596 196.484 372.596 196.003V180.468C372.595 179.988 372.396 179.527 372.043 179.187C371.69 178.847 371.212 178.655 370.713 178.654V178.654Z" fill="white"/>
-<path id="Key_z" d="M370.713 204.475H354.591C354.092 204.475 353.613 204.666 353.26 205.007C352.907 205.347 352.709 205.808 352.708 206.289V221.824C352.708 222.062 352.756 222.298 352.851 222.519C352.945 222.739 353.084 222.939 353.259 223.108C353.434 223.277 353.642 223.41 353.87 223.501C354.099 223.592 354.344 223.639 354.591 223.638H370.713C371.213 223.638 371.691 223.447 372.045 223.107C372.398 222.766 372.596 222.305 372.596 221.824V206.289C372.595 205.809 372.396 205.348 372.043 205.008C371.69 204.668 371.212 204.476 370.713 204.475V204.475Z" fill="white"/>
-<path id="Vector_11" d="M377.056 231.587H360.935C360.687 231.586 360.442 231.633 360.213 231.724C359.985 231.815 359.777 231.948 359.602 232.117C359.427 232.286 359.288 232.486 359.194 232.706C359.099 232.927 359.051 233.163 359.052 233.401V248.936C359.051 249.174 359.099 249.41 359.194 249.63C359.288 249.851 359.427 250.051 359.602 250.22C359.777 250.388 359.985 250.522 360.213 250.613C360.442 250.704 360.687 250.751 360.935 250.75H377.056C377.556 250.75 378.034 250.559 378.388 250.218C378.741 249.878 378.939 249.417 378.939 248.936V233.401C378.939 232.92 378.741 232.459 378.388 232.118C378.034 231.778 377.556 231.587 377.056 231.587V231.587Z" fill="#E6E6E6"/>
-<path id="Vector_12" d="M410.922 125.651H394.8C394.301 125.651 393.822 125.842 393.469 126.182C393.116 126.522 392.918 126.984 392.917 127.465V143.006C392.918 143.244 392.966 143.48 393.061 143.699C393.156 143.919 393.295 144.119 393.47 144.286C393.645 144.454 393.852 144.587 394.081 144.678C394.309 144.768 394.553 144.815 394.8 144.814H410.922C411.42 144.814 411.898 144.624 412.251 144.285C412.604 143.946 412.803 143.486 412.805 143.006V127.465C412.804 126.984 412.605 126.523 412.252 126.183C411.899 125.844 411.421 125.652 410.922 125.651V125.651Z" fill="white"/>
-<path id="Key_t" d="M448.208 151.437H432.086C431.587 151.437 431.108 151.628 430.755 151.969C430.402 152.309 430.203 152.77 430.203 153.251V168.793C430.203 169.031 430.252 169.266 430.347 169.486C430.442 169.706 430.581 169.905 430.755 170.073C430.93 170.241 431.138 170.374 431.366 170.464C431.595 170.555 431.839 170.601 432.086 170.6H448.208C448.706 170.601 449.184 170.41 449.537 170.072C449.89 169.733 450.089 169.273 450.091 168.793V153.251C450.089 152.771 449.891 152.31 449.538 151.97C449.185 151.63 448.707 151.438 448.208 151.437V151.437Z" fill="white"/>
-<path id="Key_d" d="M397.51 179.352H381.389C380.889 179.352 380.411 179.543 380.057 179.883C379.704 180.224 379.506 180.685 379.506 181.166V196.701C379.505 196.939 379.554 197.175 379.648 197.396C379.742 197.616 379.881 197.816 380.056 197.985C380.231 198.153 380.439 198.287 380.668 198.378C380.896 198.469 381.141 198.516 381.389 198.515H397.51C398.01 198.515 398.489 198.324 398.842 197.983C399.195 197.643 399.393 197.182 399.394 196.701V181.166C399.392 180.685 399.193 180.225 398.84 179.885C398.488 179.545 398.009 179.353 397.51 179.352V179.352Z" fill="white"/>
-<path id="Key_x" d="M400.407 204.475H384.286C383.786 204.475 383.307 204.666 382.954 205.007C382.601 205.347 382.403 205.808 382.403 206.289V221.824C382.402 222.062 382.451 222.298 382.545 222.519C382.639 222.739 382.778 222.939 382.953 223.108C383.128 223.277 383.336 223.41 383.564 223.501C383.793 223.592 384.038 223.639 384.286 223.638H400.407C400.907 223.638 401.386 223.447 401.739 223.107C402.092 222.766 402.29 222.305 402.291 221.824V206.289C402.289 205.809 402.09 205.348 401.737 205.008C401.385 204.668 400.906 204.476 400.407 204.475V204.475Z" fill="white"/>
-<path id="Vector_13" d="M405.477 231.692H389.355C389.108 231.691 388.863 231.738 388.634 231.829C388.406 231.92 388.198 232.053 388.023 232.222C387.848 232.391 387.709 232.591 387.615 232.811C387.52 233.032 387.472 233.268 387.472 233.506V249.041C387.472 249.279 387.52 249.515 387.615 249.735C387.709 249.956 387.848 250.156 388.023 250.325C388.198 250.493 388.406 250.627 388.634 250.718C388.863 250.809 389.108 250.856 389.355 250.855H405.477C405.977 250.855 406.455 250.664 406.808 250.323C407.162 249.983 407.36 249.522 407.36 249.041V233.506C407.36 233.025 407.162 232.564 406.808 232.223C406.455 231.883 405.977 231.692 405.477 231.692V231.692Z" fill="#E6E6E6"/>
-<path id="Vector_14" d="M444.788 125.651H428.659C428.161 125.652 427.683 125.844 427.332 126.184C426.98 126.524 426.783 126.985 426.783 127.465V143.006C426.783 143.486 426.981 143.945 427.333 144.284C427.684 144.623 428.161 144.814 428.659 144.814H444.788C445.286 144.814 445.764 144.624 446.117 144.285C446.47 143.946 446.669 143.486 446.671 143.006V127.465C446.67 126.984 446.471 126.523 446.118 126.183C445.765 125.844 445.287 125.652 444.788 125.651Z" fill="white"/>
-<path id="Key_y" d="M477.178 150.739H461.049C460.551 150.741 460.073 150.933 459.722 151.273C459.37 151.613 459.173 152.074 459.173 152.554V168.095C459.173 168.574 459.371 169.034 459.722 169.373C460.074 169.712 460.551 169.902 461.049 169.903H477.178C477.676 169.903 478.154 169.713 478.507 169.374C478.86 169.035 479.059 168.575 479.061 168.095V152.554C479.059 152.073 478.861 151.612 478.508 151.272C478.155 150.932 477.677 150.741 477.178 150.739V150.739Z" fill="white"/>
-<path id="Key_g" d="M451.105 178.654H434.976C434.478 178.655 434 178.847 433.649 179.188C433.297 179.528 433.1 179.988 433.1 180.468V196.003C433.099 196.241 433.147 196.476 433.241 196.696C433.335 196.916 433.473 197.116 433.647 197.285C433.822 197.453 434.029 197.587 434.257 197.678C434.485 197.77 434.729 197.817 434.976 197.817H451.105C451.604 197.817 452.083 197.626 452.436 197.286C452.789 196.945 452.988 196.484 452.988 196.003V180.468C452.986 179.988 452.788 179.527 452.435 179.187C452.082 178.847 451.604 178.655 451.105 178.654V178.654Z" fill="white"/>
-<path id="Key_f" d="M424.308 178.654H408.179C407.68 178.655 407.203 178.847 406.852 179.188C406.5 179.528 406.303 179.988 406.303 180.468V196.003C406.302 196.241 406.35 196.476 406.444 196.696C406.538 196.916 406.676 197.116 406.85 197.285C407.025 197.453 407.232 197.587 407.459 197.678C407.687 197.77 407.932 197.817 408.179 197.817H424.308C424.807 197.817 425.286 197.626 425.639 197.286C425.992 196.945 426.191 196.484 426.191 196.003V180.468C426.189 179.988 425.99 179.527 425.638 179.187C425.285 178.847 424.807 178.655 424.308 178.654V178.654Z" fill="white"/>
-<path id="Key_h" d="M478.626 178.654H462.497C461.999 178.655 461.522 178.847 461.17 179.188C460.819 179.528 460.621 179.988 460.621 180.468V196.003C460.621 196.241 460.669 196.476 460.763 196.696C460.857 196.916 460.995 197.116 461.169 197.285C461.343 197.453 461.55 197.587 461.778 197.678C462.006 197.77 462.25 197.817 462.497 197.817H478.626C479.126 197.817 479.604 197.626 479.957 197.286C480.311 196.945 480.509 196.484 480.509 196.003V180.468C480.508 179.988 480.309 179.527 479.956 179.187C479.603 178.847 479.125 178.655 478.626 178.654V178.654Z" fill="white"/>
-<path id="Key_v" d="M458.347 204.475H442.218C441.72 204.477 441.243 204.669 440.891 205.009C440.54 205.349 440.342 205.809 440.342 206.289V221.824C440.342 222.062 440.39 222.297 440.484 222.517C440.578 222.737 440.716 222.937 440.89 223.106C441.064 223.274 441.271 223.408 441.499 223.5C441.727 223.591 441.971 223.638 442.218 223.638H458.347C458.847 223.638 459.325 223.447 459.679 223.107C460.032 222.766 460.23 222.305 460.23 221.824V206.289C460.229 205.809 460.03 205.348 459.677 205.008C459.324 204.668 458.846 204.476 458.347 204.475V204.475Z" fill="white"/>
-<path id="Key_c" d="M428.653 205.173H412.524C412.026 205.174 411.549 205.366 411.197 205.707C410.846 206.047 410.648 206.507 410.648 206.987V222.522C410.648 222.76 410.696 222.995 410.79 223.215C410.883 223.435 411.022 223.635 411.196 223.804C411.37 223.972 411.577 224.106 411.805 224.197C412.033 224.289 412.277 224.336 412.524 224.336H428.653C429.152 224.336 429.631 224.145 429.984 223.805C430.338 223.464 430.536 223.003 430.536 222.522V206.987C430.535 206.506 430.336 206.046 429.983 205.706C429.63 205.366 429.152 205.174 428.653 205.173V205.173Z" fill="white"/>
-<path id="Vector_15" d="M478.654 125.651H462.525C462.026 125.652 461.549 125.844 461.198 126.184C460.846 126.524 460.649 126.985 460.649 127.465V143.006C460.649 143.486 460.847 143.945 461.198 144.284C461.55 144.623 462.027 144.814 462.525 144.814H478.654C479.152 144.814 479.63 144.624 479.983 144.285C480.336 143.946 480.535 143.486 480.537 143.006V127.465C480.535 126.984 480.336 126.523 479.984 126.184C479.631 125.844 479.153 125.652 478.654 125.651V125.651Z" fill="white"/>
-<path id="Key_u" d="M505.423 152.135H489.294C488.796 152.137 488.319 152.328 487.967 152.669C487.616 153.009 487.418 153.469 487.419 153.949V169.491C487.419 169.97 487.616 170.43 487.968 170.769C488.32 171.108 488.797 171.298 489.294 171.298H505.423C505.922 171.298 506.4 171.108 506.753 170.769C507.105 170.431 507.305 169.971 507.306 169.491V153.949C507.305 153.469 507.106 153.008 506.753 152.668C506.4 152.328 505.922 152.136 505.423 152.135V152.135Z" fill="white"/>
-<path id="Key_j" d="M505.423 179.352H489.294C488.796 179.353 488.319 179.545 487.967 179.885C487.616 180.226 487.418 180.686 487.419 181.166V196.701C487.418 196.939 487.466 197.174 487.56 197.394C487.654 197.614 487.792 197.814 487.966 197.983C488.14 198.151 488.347 198.285 488.575 198.376C488.803 198.468 489.048 198.515 489.294 198.515H505.423C505.923 198.515 506.401 198.324 506.755 197.983C507.108 197.643 507.306 197.182 507.306 196.701V181.166C507.305 180.685 507.106 180.225 506.753 179.885C506.4 179.545 505.922 179.353 505.423 179.352V179.352Z" fill="white"/>
-<path id="Key_b" d="M485.869 205.173H469.74C469.241 205.174 468.764 205.366 468.413 205.707C468.061 206.047 467.864 206.507 467.864 206.987V222.522C467.863 222.76 467.911 222.995 468.005 223.215C468.099 223.435 468.237 223.635 468.411 223.804C468.586 223.972 468.793 224.106 469.02 224.197C469.248 224.289 469.493 224.336 469.74 224.336H485.869C486.368 224.336 486.847 224.145 487.2 223.805C487.553 223.464 487.752 223.003 487.752 222.522V206.987C487.75 206.506 487.551 206.046 487.199 205.706C486.846 205.366 486.368 205.174 485.869 205.173V205.173Z" fill="white"/>
-<path id="Vector_16" d="M512.519 125.651H496.391C495.892 125.652 495.415 125.844 495.064 126.184C494.712 126.524 494.515 126.985 494.515 127.465V143.006C494.515 143.486 494.712 143.945 495.064 144.284C495.416 144.623 495.893 144.814 496.391 144.814H512.519C513.018 144.814 513.496 144.624 513.849 144.285C514.202 143.946 514.401 143.486 514.403 143.006V127.465C514.401 126.984 514.202 126.523 513.849 126.184C513.497 125.844 513.018 125.652 512.519 125.651V125.651Z" fill="white"/>
-<path id="Key_i" d="M533.669 152.135H517.54C517.042 152.137 516.565 152.328 516.213 152.669C515.861 153.009 515.664 153.469 515.664 153.949V169.491C515.664 169.97 515.862 170.43 516.214 170.769C516.565 171.108 517.043 171.298 517.54 171.298H533.669C534.167 171.298 534.645 171.108 534.998 170.769C535.351 170.431 535.55 169.971 535.552 169.491V153.949C535.551 153.469 535.352 153.008 534.999 152.668C534.646 152.328 534.168 152.136 533.669 152.135Z" fill="white"/>
-<path id="Key_k" d="M532.22 179.352H516.092C515.593 179.353 515.116 179.545 514.764 179.885C514.413 180.226 514.216 180.686 514.216 181.166V196.701C514.215 196.939 514.263 197.174 514.357 197.394C514.451 197.614 514.589 197.814 514.763 197.983C514.937 198.151 515.144 198.285 515.372 198.376C515.6 198.468 515.845 198.515 516.092 198.515H532.22C532.72 198.515 533.199 198.324 533.552 197.983C533.905 197.643 534.103 197.182 534.104 196.701V181.166C534.102 180.685 533.903 180.225 533.55 179.885C533.198 179.545 532.719 179.353 532.22 179.352Z" fill="white"/>
-<path id="Key_n" d="M511.942 205.173H495.813C495.314 205.174 494.837 205.366 494.486 205.707C494.134 206.047 493.937 206.507 493.937 206.987V222.522C493.936 222.76 493.984 222.995 494.078 223.215C494.172 223.435 494.31 223.635 494.484 223.804C494.658 223.972 494.865 224.106 495.093 224.197C495.321 224.289 495.566 224.336 495.813 224.336H511.942C512.441 224.336 512.92 224.145 513.273 223.805C513.626 223.464 513.824 223.003 513.825 222.522V206.987C513.823 206.506 513.624 206.046 513.271 205.706C512.919 205.366 512.441 205.174 511.942 205.173Z" fill="white"/>
-<path id="Key_Spacebar" d="M501.024 231.692H425.254C422.914 231.693 420.672 231.882 419.02 232.216C417.369 232.551 416.442 233.004 416.442 233.476V248.75C416.439 248.984 416.665 249.216 417.106 249.432C417.547 249.648 418.196 249.845 419.014 250.011C419.832 250.176 420.805 250.308 421.876 250.398C422.947 250.487 424.095 250.534 425.254 250.534H501.024C503.37 250.534 505.619 250.346 507.278 250.011C508.937 249.677 509.869 249.223 509.87 248.75V233.476C509.863 233.003 508.929 232.55 507.272 232.216C505.614 231.881 503.368 231.693 501.024 231.692V231.692Z" fill="white"/>
-<path id="Vector_17" d="M535.842 231.692H519.713C519.466 231.692 519.221 231.739 518.994 231.83C518.766 231.922 518.559 232.055 518.384 232.224C518.21 232.393 518.072 232.593 517.978 232.813C517.884 233.033 517.836 233.268 517.837 233.506V249.041C517.836 249.278 517.884 249.514 517.978 249.734C518.072 249.954 518.21 250.154 518.384 250.323C518.559 250.491 518.766 250.625 518.994 250.716C519.221 250.808 519.466 250.855 519.713 250.855H535.842C536.341 250.855 536.82 250.664 537.173 250.323C537.526 249.983 537.725 249.522 537.725 249.041V233.506C537.725 233.025 537.526 232.564 537.173 232.223C536.82 231.883 536.341 231.692 535.842 231.692Z" fill="#E6E6E6"/>
-<path id="Vector_18" d="M546.385 125.651H530.256C529.758 125.652 529.281 125.844 528.929 126.184C528.578 126.524 528.38 126.985 528.381 127.465V143.006C528.381 143.486 528.578 143.945 528.93 144.284C529.282 144.623 529.759 144.814 530.256 144.814H546.385C546.884 144.814 547.362 144.624 547.715 144.285C548.067 143.946 548.267 143.486 548.268 143.006V127.465C548.267 126.984 548.068 126.523 547.715 126.184C547.362 125.844 546.884 125.652 546.385 125.651Z" fill="#E6E6E6"/>
-<path id="Key_o" d="M560.466 152.135H544.337C543.839 152.137 543.362 152.328 543.01 152.669C542.659 153.009 542.461 153.469 542.461 153.949V169.491C542.461 169.97 542.659 170.43 543.011 170.769C543.363 171.108 543.84 171.298 544.337 171.298H560.466C560.964 171.298 561.442 171.108 561.795 170.769C562.148 170.431 562.347 169.971 562.349 169.491V153.949C562.348 153.469 562.149 153.008 561.796 152.668C561.443 152.328 560.965 152.136 560.466 152.135V152.135Z" fill="white"/>
-<path id="Key_l" d="M557.569 178.654H541.44C540.942 178.655 540.465 178.847 540.113 179.188C539.762 179.528 539.564 179.988 539.564 180.468V196.003C539.564 196.241 539.612 196.476 539.706 196.696C539.8 196.916 539.938 197.116 540.112 197.285C540.286 197.453 540.493 197.587 540.721 197.678C540.949 197.77 541.193 197.817 541.44 197.817H557.569C558.069 197.817 558.547 197.626 558.9 197.286C559.253 196.945 559.452 196.484 559.452 196.003V180.468C559.451 179.988 559.252 179.527 558.899 179.187C558.546 178.847 558.068 178.655 557.569 178.654Z" fill="white"/>
-<path id="Key_m" d="M538.014 205.871H521.886C521.387 205.872 520.91 206.064 520.559 206.404C520.207 206.744 520.01 207.205 520.01 207.685V223.22C520.009 223.457 520.057 223.693 520.151 223.913C520.245 224.133 520.383 224.333 520.557 224.502C520.731 224.67 520.938 224.804 521.166 224.895C521.394 224.987 521.639 225.034 521.886 225.034H538.014C538.514 225.034 538.993 224.843 539.346 224.502C539.699 224.162 539.897 223.701 539.898 223.22V207.685C539.896 207.204 539.697 206.744 539.344 206.404C538.992 206.064 538.513 205.872 538.014 205.871V205.871Z" fill="white"/>
-<path id="Vector_19" d="M564.812 230.994H548.683C548.436 230.994 548.191 231.041 547.963 231.133C547.736 231.224 547.528 231.358 547.354 231.526C547.18 231.695 547.042 231.895 546.948 232.115C546.854 232.335 546.806 232.57 546.807 232.808V248.343C546.806 248.581 546.854 248.816 546.948 249.036C547.042 249.256 547.18 249.456 547.354 249.625C547.528 249.793 547.735 249.927 547.963 250.018C548.191 250.11 548.436 250.157 548.683 250.157H564.812C565.311 250.157 565.79 249.966 566.143 249.626C566.496 249.285 566.694 248.824 566.695 248.343V232.808C566.694 232.327 566.496 231.866 566.143 231.525C565.79 231.185 565.311 230.994 564.812 230.994V230.994Z" fill="#E6E6E6"/>
-<path id="Vector_20" d="M574.649 126.314H558.791C558.301 126.315 557.832 126.483 557.486 126.78C557.14 127.077 556.946 127.48 556.946 127.9V141.483C556.946 141.902 557.141 142.304 557.487 142.6C557.832 142.896 558.302 143.063 558.791 143.063H574.649C575.139 143.063 575.609 142.897 575.956 142.6C576.303 142.304 576.499 141.902 576.501 141.483V127.9C576.5 127.479 576.304 127.077 575.957 126.78C575.61 126.482 575.14 126.315 574.649 126.314V126.314Z" fill="#E6E6E6"/>
-<path id="Key_p" d="M588.712 152.135H572.583C572.085 152.137 571.607 152.328 571.256 152.669C570.904 153.009 570.707 153.469 570.707 153.949V169.491C570.707 169.97 570.905 170.43 571.256 170.769C571.608 171.108 572.085 171.298 572.583 171.298H588.712C589.21 171.298 589.688 171.108 590.041 170.769C590.394 170.431 590.593 169.971 590.595 169.491V153.949C590.593 153.469 590.394 153.008 590.042 152.668C589.689 152.328 589.211 152.136 588.712 152.135V152.135Z" fill="white"/>
-<path id="Vector_21" d="M579.406 179.352H567.072C566.691 179.353 566.326 179.528 566.057 179.838C565.788 180.147 565.637 180.567 565.637 181.004V195.146C565.637 195.363 565.673 195.578 565.745 195.778C565.817 195.978 565.923 196.16 566.056 196.314C566.189 196.467 566.347 196.589 566.522 196.672C566.696 196.755 566.883 196.798 567.072 196.798H579.406C579.788 196.798 580.154 196.624 580.424 196.314C580.694 196.005 580.846 195.585 580.846 195.146V181.004C580.845 180.566 580.693 180.146 580.423 179.837C580.154 179.527 579.788 179.353 579.406 179.352V179.352Z" fill="#E6E6E6"/>
-<path id="Key_Comma" d="M564.087 205.173H547.958C547.46 205.174 546.983 205.366 546.631 205.707C546.28 206.047 546.082 206.507 546.083 206.987V222.522C546.082 222.76 546.13 222.995 546.224 223.215C546.318 223.435 546.456 223.635 546.63 223.804C546.804 223.972 547.011 224.106 547.239 224.197C547.467 224.289 547.711 224.336 547.958 224.336H564.087C564.587 224.336 565.066 224.145 565.419 223.805C565.772 223.464 565.97 223.003 565.97 222.522V206.987C565.969 206.506 565.77 206.046 565.417 205.706C565.064 205.366 564.586 205.174 564.087 205.173Z" fill="#E6E6E6"/>
-<path id="Key_Backspace" d="M613.454 125.616H588.785C588.023 125.618 587.293 125.813 586.756 126.16C586.218 126.507 585.916 126.976 585.916 127.466V143.313C585.916 143.802 586.219 144.271 586.757 144.616C587.295 144.962 588.024 145.156 588.785 145.156H613.454C614.216 145.156 614.948 144.962 615.487 144.617C616.027 144.272 616.332 143.803 616.335 143.313V127.466C616.332 126.976 616.028 126.506 615.489 126.159C614.949 125.813 614.218 125.617 613.454 125.616V125.616Z" fill="#E6E6E6"/>
-<path id="Vector_22" d="M614.117 152.135H597.988C597.49 152.136 597.013 152.328 596.661 152.668C596.309 153.008 596.112 153.469 596.112 153.949V169.491C596.112 169.97 596.31 170.43 596.662 170.768C597.013 171.107 597.491 171.298 597.988 171.298H614.117C614.615 171.298 615.093 171.108 615.446 170.769C615.799 170.43 615.998 169.971 616 169.491V153.949C615.999 153.468 615.8 153.008 615.447 152.668C615.094 152.328 614.616 152.136 614.117 152.135V152.135Z" fill="#E6E6E6"/>
-<path id="Vector_23" d="M613.454 178.654H588.785C588.023 178.655 587.293 178.844 586.756 179.179C586.218 179.513 585.916 179.966 585.916 180.438V195.712C585.915 195.946 585.989 196.178 586.132 196.394C586.276 196.61 586.487 196.807 586.754 196.973C587.02 197.138 587.337 197.27 587.685 197.36C588.034 197.45 588.408 197.496 588.785 197.496H613.454C614.218 197.496 614.95 197.308 615.491 196.973C616.031 196.639 616.334 196.185 616.335 195.712V180.438C616.332 179.965 616.028 179.512 615.489 179.178C614.949 178.844 614.218 178.655 613.454 178.654V178.654Z" fill="#E6E6E6"/>
-<path id="Key_Shift" d="M612.152 205.173H576.323C575.216 205.174 574.156 205.363 573.375 205.698C572.594 206.032 572.155 206.485 572.156 206.957V222.231C572.154 222.465 572.261 222.697 572.469 222.913C572.678 223.129 572.985 223.326 573.372 223.492C573.759 223.657 574.219 223.789 574.725 223.879C575.231 223.969 575.774 224.015 576.323 224.015H612.152C613.261 224.015 614.324 223.827 615.109 223.492C615.893 223.158 616.334 222.704 616.335 222.231V206.957C616.331 206.484 615.89 206.031 615.106 205.697C614.322 205.363 613.26 205.174 612.152 205.173V205.173Z" fill="white"/>
-<path id="Vector_24" d="M615.374 240.764H607.151C607.026 240.764 606.901 240.786 606.785 240.83C606.669 240.873 606.563 240.936 606.474 241.016C606.385 241.096 606.315 241.19 606.267 241.295C606.219 241.399 606.195 241.51 606.195 241.623V248.977C606.195 249.09 606.219 249.201 606.267 249.306C606.315 249.41 606.385 249.504 606.474 249.584C606.563 249.664 606.669 249.727 606.785 249.77C606.901 249.814 607.026 249.836 607.151 249.836H615.374C615.629 249.836 615.873 249.746 616.053 249.584C616.233 249.423 616.334 249.205 616.335 248.977V241.623C616.334 241.395 616.233 241.177 616.053 241.016C615.873 240.855 615.629 240.764 615.374 240.764V240.764Z" fill="#E6E6E6"/>
-<path id="Vector_25" d="M600.165 240.764H591.942C591.816 240.764 591.692 240.786 591.576 240.83C591.459 240.873 591.354 240.936 591.265 241.016C591.176 241.096 591.106 241.19 591.058 241.295C591.01 241.399 590.986 241.51 590.986 241.623V248.977C590.986 249.09 591.01 249.201 591.058 249.306C591.106 249.41 591.176 249.504 591.265 249.584C591.354 249.664 591.459 249.727 591.576 249.77C591.692 249.814 591.816 249.836 591.942 249.836H600.165C600.42 249.836 600.664 249.746 600.844 249.584C601.024 249.423 601.125 249.205 601.125 248.977V241.623C601.125 241.395 601.024 241.177 600.844 241.016C600.664 240.855 600.42 240.764 600.165 240.764V240.764Z" fill="#E6E6E6"/>
-<path id="Vector_26" d="M600.165 228.202H591.942C591.816 228.202 591.692 228.225 591.576 228.268C591.459 228.311 591.354 228.375 591.265 228.454C591.176 228.534 591.106 228.629 591.058 228.733C591.01 228.837 590.986 228.949 590.986 229.061V236.416C590.986 236.528 591.01 236.64 591.058 236.744C591.106 236.848 591.176 236.943 591.265 237.022C591.354 237.102 591.459 237.166 591.576 237.209C591.692 237.252 591.816 237.274 591.942 237.275H600.165C600.42 237.274 600.664 237.184 600.844 237.023C601.024 236.862 601.125 236.643 601.125 236.416V229.061C601.125 228.834 601.024 228.615 600.844 228.454C600.664 228.293 600.42 228.202 600.165 228.202V228.202Z" fill="#E6E6E6"/>
-<path id="Vector_27" d="M584.956 241.462H576.733C576.607 241.462 576.483 241.484 576.366 241.527C576.25 241.571 576.145 241.634 576.056 241.714C575.967 241.794 575.897 241.888 575.849 241.992C575.801 242.097 575.776 242.208 575.777 242.321V249.675C575.776 249.788 575.801 249.899 575.849 250.003C575.897 250.107 575.967 250.202 576.056 250.282C576.145 250.362 576.25 250.425 576.366 250.468C576.483 250.512 576.607 250.534 576.733 250.534H584.956C585.211 250.534 585.455 250.443 585.635 250.282C585.815 250.121 585.916 249.903 585.916 249.675V242.321C585.916 242.093 585.815 241.875 585.635 241.713C585.455 241.552 585.211 241.462 584.956 241.462V241.462Z" fill="#E6E6E6"/>
-<path id="Vector_28" d="M243.417 460.627H242.967L243.351 460.579L243.417 460.627Z" fill="#2F2E41"/>
-<path id="Pant" d="M303.878 516.937C301.959 516.979 300.033 517 298.099 517C252.16 517 194.379 493.356 141.944 460.626C137.606 457.919 133.306 455.148 129.045 452.315L132.782 430.402L245.917 415.928L302.698 457.765L303.878 516.937Z" fill="#2F2E41"/>
-<path id="Shirt" d="M332.826 240.84L332.667 241.65L330.878 250.75L318.515 313.698L317.784 317.403L316.248 325.24L314.148 335.932L309.491 359.624L308.159 366.414L307.659 368.962L307.485 369.855L306.102 376.882L305.682 379.004L302.307 403.101L299.359 424.1L309.223 433.605L303.342 439.272L299.482 460.578L299.489 460.626H141.944C135.57 456.649 129.289 452.547 123.099 448.323L124.395 438.972C124.395 438.972 128.857 422.272 126.981 420.276C125.112 418.28 121.013 414.714 121.013 414.714L126.177 403.101L126.894 401.489L125.692 377.462L125.663 376.882L125.264 368.962L125.214 367.943L125.134 366.414L123.606 335.932L123.577 335.429L123.113 326.189L122.679 317.403L122.491 313.698L120.629 276.62L119.369 251.441L118.268 229.521L122.693 228.076L170.023 212.577L189.628 188.019L230.432 188.312L250.407 188.458L283.737 218.243L291.32 221.733L291.378 221.761L296.817 224.266L310.607 230.61L325.186 237.323L332.667 240.764L332.826 240.84Z" fill="#CACACA"/>
-<g id="Left Hand">
-<path id="Vector_29" d="M194.901 389.096C191.911 390.212 188.704 390.683 185.504 390.476C182.305 390.268 179.191 389.387 176.384 387.895C173.576 386.402 171.142 384.335 169.253 381.838C167.364 379.341 166.066 376.475 165.451 373.443L88.2757 362.893L115.104 333.651L184.277 348.152C189.633 347.494 195.053 348.752 199.511 351.687C203.968 354.622 207.153 359.03 208.463 364.077C209.772 369.124 209.115 374.459 206.616 379.07C204.116 383.681 199.948 387.248 194.901 389.096V389.096Z" fill="#FFB6B6"/>
-<path id="Vector_30" d="M160.492 335.932C159.062 334.185 157.038 332.982 154.778 332.533L123.113 326.189L119.862 325.54C117.972 325.16 116.186 324.4 114.623 323.309C113.059 322.219 111.752 320.822 110.787 319.211C109.786 317.519 109.18 315.638 109.012 313.698C108.861 311.934 109.081 310.158 109.657 308.478L120.629 276.62L127.748 255.949C128.611 253.456 129.121 250.862 129.262 248.238C129.659 240.975 127.327 233.818 122.693 228.076C118.62 223.03 112.97 219.381 106.578 217.666C100.186 215.952 93.3913 216.265 87.1979 218.557C87.0169 218.62 86.8285 218.69 86.6475 218.76C82.2059 220.469 78.2376 223.15 75.0562 226.591C71.8749 230.032 69.5676 234.138 68.3168 238.586L60.5818 265.698L44.6266 321.591C43.0295 327.19 42.7068 333.057 43.6807 338.788C44.6546 344.519 46.9022 349.98 50.2695 354.797C53.6368 359.613 58.0443 363.672 63.1905 366.695C68.3367 369.718 74.0999 371.634 80.0858 372.311L125.692 377.462L146.796 379.848C149.237 380.131 151.698 379.517 153.69 378.129C155.682 376.742 157.058 374.682 157.544 372.36C157.578 372.234 157.605 372.106 157.624 371.976L158.131 368.962L162.426 343.573L162.469 343.315C162.693 342.019 162.632 340.693 162.292 339.422C161.951 338.15 161.338 336.961 160.492 335.932Z" fill="#CACACA"/>
-</g>
-<g id="Right Hand">
-<path id="Vector_31" d="M244.059 389.91C247.182 390.609 250.427 390.639 253.564 389.999C256.701 389.358 259.654 388.063 262.216 386.204C264.777 384.345 266.884 381.968 268.389 379.239C269.894 376.51 270.759 373.497 270.924 370.411L345.784 349.48L314.944 324.175L248.584 347.932C243.186 348.008 238.004 349.989 234.021 353.501C230.037 357.012 227.529 361.81 226.971 366.984C226.413 372.158 227.845 377.35 230.994 381.576C234.143 385.801 238.792 388.766 244.059 389.91V389.91Z" fill="#FFB6B6"/>
-<path id="Vector_32" d="M382.952 302.657L347.348 223.701C344.435 217.218 339.234 211.94 332.667 208.801C330.637 207.838 328.505 207.09 326.308 206.568C325.968 206.484 325.627 206.408 325.287 206.345C320.629 205.365 315.797 205.451 311.179 206.596C306.958 207.641 303.012 209.532 299.601 212.145C296.189 214.757 293.387 218.034 291.378 221.761C291.356 221.792 291.336 221.824 291.32 221.858C289.076 226.042 287.882 230.675 287.836 235.384C287.789 240.093 288.891 244.747 291.052 248.97L291.494 249.822L316.654 298.512C317.513 300.177 317.997 301.998 318.075 303.856C318.152 305.714 317.821 307.567 317.102 309.293C316.384 311.019 315.296 312.579 313.908 313.871C312.52 315.163 310.864 316.157 309.049 316.789L307.282 317.403L275.502 328.464C273.118 329.284 271.156 330.964 270.031 333.149C268.906 335.335 268.704 337.856 269.469 340.182L269.549 340.426L278.457 367.894C278.501 368.019 278.544 368.145 278.595 368.271C278.673 368.508 278.77 368.739 278.884 368.962C279.831 371.008 281.505 372.661 283.604 373.624C285.703 374.588 288.089 374.798 290.335 374.217L307.485 369.855L310.983 368.962L355.271 357.691C361.098 356.207 366.524 353.528 371.177 349.836C375.83 346.143 379.599 341.526 382.228 336.3C384.857 331.073 386.283 325.361 386.407 319.555C386.532 313.748 385.353 307.984 382.952 302.657L382.952 302.657Z" fill="#CACACA"/>
-</g>
-<path id="Vector_33" d="M483.818 440.66C481.044 443.459 478.195 446.19 475.271 448.853H405.592V440.66H483.818Z" fill="white"/>
-<path id="Table" d="M32.5173 368.962C42.7802 381.023 53.8484 392.427 65.6516 403.101H392.99C396.143 403.106 399.166 404.316 401.396 406.465C403.625 408.615 404.879 411.528 404.882 414.567L405.353 494.494C425.49 485.579 444.348 474.196 461.46 460.626C466.211 456.865 470.815 452.941 475.271 448.853C478.19 446.188 481.039 443.457 483.818 440.66C504.65 419.749 521.519 395.483 533.66 368.962H32.5173Z" fill="#3F3D56"/>
-<g id="Monitor">
-<path id="Vector_34" d="M332.175 178.619C331.83 178.006 331.321 177.495 330.7 177.139C330.079 176.783 329.37 176.595 328.648 176.595H91.2103C90.1458 176.597 89.1257 177.007 88.3737 177.733C87.6216 178.459 87.1988 179.442 87.198 180.468V317.403H332.667V180.468C332.671 179.821 332.501 179.184 332.175 178.619V178.619Z" fill="#3F3D56"/>
-<path id="Vector_35" d="M86.9829 313.696V330.594C86.9829 331.294 87.1261 331.988 87.4044 332.635C87.6826 333.283 88.0903 333.871 88.6044 334.366C89.1185 334.861 89.7288 335.254 90.4004 335.522C91.0721 335.79 91.792 335.928 92.519 335.928H183.12V369.277H181.916C181.822 369.277 181.728 369.295 181.641 369.33C181.554 369.365 181.475 369.416 181.408 369.48C181.341 369.544 181.288 369.621 181.252 369.705C181.216 369.789 181.197 369.879 181.197 369.97V372.289C181.197 372.38 181.216 372.471 181.252 372.555C181.288 372.639 181.341 372.715 181.408 372.78C181.475 372.844 181.554 372.895 181.641 372.93C181.728 372.965 181.822 372.982 181.916 372.982H238.16C238.255 372.982 238.348 372.965 238.435 372.93C238.523 372.895 238.602 372.844 238.669 372.78C238.736 372.715 238.789 372.639 238.825 372.555C238.861 372.471 238.879 372.38 238.879 372.289V369.97C238.879 369.879 238.861 369.789 238.825 369.705C238.789 369.621 238.736 369.544 238.669 369.48C238.602 369.416 238.523 369.365 238.435 369.33C238.348 369.295 238.255 369.277 238.16 369.277H236.957V335.928H327.558C328.285 335.928 329.005 335.79 329.676 335.522C330.348 335.254 330.958 334.861 331.472 334.366C331.986 333.871 332.394 333.283 332.672 332.635C332.95 331.988 333.094 331.294 333.094 330.594V313.696H86.9829Z" fill="#3F3D56"/>
-</g>
-<path id="Monitor_Logo" d="M207.064 265.196C219.587 265.196 229.738 255.415 229.738 243.349C229.738 231.282 219.587 221.501 207.064 221.501C194.542 221.501 184.391 231.282 184.391 243.349C184.391 255.415 194.542 265.196 207.064 265.196Z" fill="#6C63FF"/>
-<g id="Head">
-<path id="ab6171fa-7d69-4734-b81c-8dff60f9761b" d="M204.828 171.015C232.156 171.015 254.31 149.669 254.31 123.336C254.31 97.0032 232.156 75.6563 204.828 75.6563C177.499 75.6563 155.346 97.0032 155.346 123.336C155.346 149.669 177.499 171.015 204.828 171.015Z" fill="#FFB6B6"/>
-<path id="bf427902-b9bf-4946-b5d7-5c1c7e04535e" d="M248.16 88.378C248.16 88.378 261.102 64.5383 232.63 62.371C232.63 62.371 208.357 41.1543 183.616 58.4923C183.616 58.4923 170.121 58.4923 162.741 73.2072C162.741 73.2072 152.127 69.3285 149.795 79.709C149.795 79.709 142.028 101.382 149.795 120.887C157.561 140.392 160.141 142.559 160.141 142.559C160.141 142.559 147.379 101.665 178.443 99.4982C209.506 97.3309 244.267 78.6254 246.856 102.465C249.445 126.305 253.34 132.179 253.34 132.179C253.34 132.179 277.928 98.1307 248.16 88.378Z" fill="#2F2E41"/>
-<path id="Vector_36" d="M252.493 103.765C251.828 104.077 251.25 104.536 250.805 105.105C250.425 105.584 250.146 106.131 249.986 106.715C249.825 107.298 249.786 107.906 249.871 108.504L253.166 131.91C253.302 132.864 253.746 133.753 254.435 134.451C255.124 135.148 256.023 135.619 257.005 135.798L252.493 103.765Z" fill="white"/>
-<path id="Vector_37" d="M270.867 122.664L269.44 112.503C269.088 109.937 267.806 107.573 265.819 105.831L269.1 129.119C269.149 129.493 269.249 129.859 269.397 130.208C270.754 127.924 271.272 125.266 270.867 122.664V122.664Z" fill="#3F3D56"/>
-<path id="Vector_38" d="M259.591 129.985C260.582 129.855 261.08 127.583 260.703 124.909C260.327 122.236 259.219 120.174 258.228 120.303C257.237 120.433 256.739 122.705 257.115 125.379C257.491 128.052 258.6 130.114 259.591 129.985Z" fill="#3F3D56"/>
-<path id="Vector_39" d="M257.418 117.423C258.409 117.294 258.907 115.021 258.531 112.348C258.154 109.674 257.046 107.612 256.055 107.742C255.064 107.871 254.566 110.143 254.942 112.817C255.319 115.49 256.427 117.553 257.418 117.423Z" fill="#3F3D56"/>
-<path id="Vector_40" d="M271.179 122.308C272.17 122.179 272.668 119.906 272.291 117.233C271.915 114.56 270.807 112.497 269.816 112.627C268.825 112.756 268.327 115.029 268.703 117.702C269.079 120.375 270.188 122.438 271.179 122.308Z" fill="#3F3D56"/>
-<path id="Vector_41" d="M206.786 77.0511C202.81 77.0399 198.843 77.4116 194.944 78.1607C190.555 79.0051 186 80.017 181.524 81.336C174.535 81.008 169.393 79.2633 168.357 76.5765C176.505 72.0404 186.449 69.1164 197.841 67.6927C202.582 67.085 207.375 66.9447 212.145 67.274C212.65 67.5592 213.087 67.9443 213.426 68.4037C213.766 68.8632 214.001 69.3867 214.115 69.9398C214.455 72.3684 211.58 74.9505 206.786 77.0511Z" fill="#3F3D56"/>
-<g id="Headphone">
-<path id="Vector_42" d="M263.156 120.625H259.534C259.534 94.2457 232.44 76.9595 205.107 76.9456C201.132 76.9344 197.165 77.3061 193.266 78.0552C188.877 78.8996 184.321 79.9115 179.845 81.2305C162.196 86.4435 145.828 96.5207 145.828 120.625C145.824 125.306 146.442 129.969 147.667 134.499L144.162 135.378C142.862 130.56 142.204 125.603 142.206 120.625C142.206 99.5634 151.18 85.0966 166.679 76.471C174.826 71.9349 184.77 69.0108 196.163 67.5872C200.904 66.9795 205.696 66.8392 210.467 67.1685C241.899 69.3458 263.156 92.5708 263.156 120.625Z" fill="#6C63FF"/>
-<path id="Vector_43" d="M269.189 122.558L267.762 112.397C267.41 109.831 266.127 107.468 264.141 105.726C262.865 104.59 261.339 103.748 259.676 103.265C258.014 102.781 256.259 102.668 254.544 102.934L252.277 103.234C251.657 103.313 251.059 103.509 250.518 103.813C249.978 104.116 249.505 104.519 249.127 105C248.746 105.479 248.467 106.026 248.307 106.609C248.147 107.192 248.108 107.801 248.193 108.398L251.488 131.805C251.624 132.759 252.068 133.648 252.756 134.345C253.445 135.043 254.344 135.514 255.327 135.692C255.617 135.748 255.913 135.776 256.21 135.776C256.423 135.776 256.636 135.762 256.847 135.734L259.382 135.406C260.928 135.208 262.415 134.708 263.753 133.937C265.09 133.165 266.25 132.138 267.161 130.919C267.356 130.653 267.545 130.381 267.718 130.102C269.076 127.819 269.594 125.161 269.189 122.558Z" fill="#6C63FF"/>
-<path id="Vector_44" d="M158.871 144.841L153.295 122.502C153.096 121.696 152.654 120.965 152.025 120.4C151.397 119.835 150.609 119.462 149.76 119.327C149.521 119.291 149.279 119.272 149.036 119.271C148.694 119.272 148.354 119.312 148.022 119.39L145.864 119.892C143.085 120.54 140.686 122.223 139.194 124.573C137.702 126.923 137.238 129.747 137.904 132.426L140.403 142.447C140.973 144.807 142.332 146.92 144.27 148.463C144.585 148.707 144.913 148.933 145.255 149.14C146.469 149.873 147.824 150.364 149.238 150.585C150.652 150.805 152.098 150.751 153.49 150.424L155.641 149.922C156.21 149.794 156.745 149.554 157.213 149.217C157.908 148.728 158.437 148.051 158.733 147.272C159.028 146.493 159.076 145.647 158.871 144.841Z" fill="#6C63FF"/>
-</g>
-<path id="Vector_45" d="M148.859 140.79C149.831 140.565 150.088 138.256 149.433 135.632C148.778 133.009 147.459 131.066 146.486 131.291C145.514 131.517 145.257 133.826 145.912 136.449C146.567 139.072 147.886 141.016 148.859 140.79Z" fill="#3F3D56"/>
-<path id="Vector_46" d="M141.616 142.884C142.588 142.658 142.846 140.349 142.19 137.726C141.535 135.103 140.216 133.159 139.244 133.385C138.272 133.61 138.015 135.919 138.67 138.542C139.325 141.165 140.644 143.109 141.616 142.884Z" fill="#3F3D56"/>
-<path id="Vector_47" d="M160.905 144.863L155.328 122.524C155.111 121.652 154.631 120.861 153.95 120.25C153.268 119.64 152.415 119.236 151.497 119.091L159.108 149.601C159.86 149.071 160.432 148.338 160.752 147.494C161.071 146.651 161.125 145.735 160.905 144.863V144.863Z" fill="#6C63FF"/>
-<path id="Vector_48" d="M140.843 128.107C140.689 127.486 140.4 126.903 139.996 126.397C139.212 128.354 139.068 130.494 139.583 132.531L142.081 142.553C142.652 144.912 144.011 147.026 145.949 148.568L140.843 128.107Z" fill="#3F3D56"/>
-</g>
-</g>
-<defs>
-<clipPath id="clip0_1_36">
-<rect width="616" height="517" fill="white"/>
-</clipPath>
-</defs>
-</svg>
-`)
+import { DEVELOPER, CANVAS } from "./constants/images.js";
+import view from './src/view.js';
 
-let canvasElm = document.getElementsByClassName('canvas-img')[0];
-canvasElm.insertAdjacentHTML('beforeend', `
-<svg width="551" height="568" viewBox="0 0 551 568" fill="none" xmlns="http://www.w3.org/2000/svg">
-<g clip-path="url(#clip0_1_2)">
-<path d="M287.004 567C286.905 567 286.806 566.998 286.707 566.996C284.191 566.876 281.827 565.947 280.107 564.404C278.388 562.86 277.445 560.821 277.476 558.71C277.476 545.867 266.522 538.376 255.667 536.648C244.811 534.922 231.339 538.527 225.717 550.476L220.605 561.223C216.185 564.916 210.197 566.99 203.954 566.99C197.711 566.99 191.722 564.916 187.303 561.223L187 560.971L187.188 560.648L277.476 405.937V402.146C277.351 400.977 277.352 399.801 277.478 398.633L277.476 8.97637C277.476 7.90829 277.732 6.85108 278.229 5.86746C278.726 4.88383 279.454 3.99383 280.37 3.25026C281.285 2.50669 282.37 1.9247 283.559 1.5388C284.748 1.1529 286.018 0.970948 287.292 1.00377C289.809 1.12422 292.173 2.05277 293.892 3.59614C295.612 5.13951 296.554 7.1788 296.524 9.28955V398.678C296.649 399.846 296.648 401.022 296.522 402.191L296.524 405.936L387 560.971L386.697 561.223C382.278 564.916 376.289 566.99 370.046 566.99C363.803 566.99 357.815 564.916 353.395 561.223L353.261 561.053L348.283 550.476C342.661 538.528 329.191 534.925 318.334 536.648C307.478 538.376 296.524 545.867 296.524 558.71V559.024C296.524 561.139 295.521 563.167 293.736 564.663C291.95 566.158 289.529 566.999 287.004 567L287.004 567Z" fill="#E6E6E6"/>
-<path d="M530.08 385H52.9211C47.3745 384.995 42.0569 383.197 38.1348 380C34.2127 376.804 32.0064 372.47 32 367.949V96.0513C32.0062 91.5305 34.2123 87.1964 38.1344 83.9998C42.0566 80.8031 47.3744 79.005 52.9211 79H530.08C535.626 79.0052 540.944 80.8034 544.866 84C548.788 87.1966 550.994 91.5307 551 96.0513V367.949C550.994 372.47 548.787 376.804 544.865 380C540.943 383.197 535.626 384.995 530.08 385Z" fill="white"/>
-<path d="M527.403 385H48.5976C42.3416 384.994 36.3438 382.949 31.9199 379.312C27.496 375.676 25.0074 370.745 25 365.602V90.3988C25.0072 85.2556 27.4958 80.3248 31.9197 76.6881C36.3436 73.0514 42.3414 71.0058 48.5976 71H527.403C533.659 71.0059 539.657 73.0516 544.081 76.6883C548.504 80.325 550.993 85.2557 551 90.3988V365.602C550.993 370.745 548.504 375.675 544.081 379.312C539.657 382.949 533.659 384.994 527.403 385V385ZM48.5976 75.2816C43.7222 75.286 39.048 76.8801 35.6005 79.7141C32.153 82.5482 30.2137 86.3907 30.2081 90.3988V365.602C30.214 369.61 32.1535 373.452 35.6009 376.286C39.0484 379.12 43.7224 380.714 48.5976 380.718H527.403C532.278 380.714 536.952 379.12 540.4 376.286C543.847 373.452 545.786 369.61 545.792 365.602V90.3988C545.786 86.3908 543.847 82.5483 540.4 79.7142C536.953 76.8801 532.279 75.286 527.403 75.2816H48.5976Z" fill="#F2F2F2"/>
-<path d="M335.706 81H238.294C232.118 80.9941 226.197 78.9227 221.83 75.2402C217.463 71.5578 215.007 66.5649 215 61.3571V61H359V61.3571C358.993 66.5649 356.537 71.5578 352.17 75.2402C347.803 78.9227 341.882 80.9941 335.706 81Z" fill="#E6E6E6"/>
-<path d="M335.706 400H238.294C232.118 399.994 226.197 397.923 221.83 394.24C217.463 390.558 215.007 385.565 215 380.357V380H359V380.357C358.993 385.565 356.537 390.558 352.17 394.24C347.803 397.923 341.882 399.994 335.706 400V400Z" fill="#E6E6E6"/>
-<path d="M437.159 156.414C452.159 156.414 464.319 144.681 464.319 130.207C464.319 115.733 452.159 104 437.159 104C422.16 104 410 115.733 410 130.207C410 144.681 422.16 156.414 437.159 156.414Z" fill="#FF6584"/>
-<path d="M397.485 118.261C384.322 116.544 369.295 123.409 365.884 135.795C365.024 133.593 363.461 131.712 361.42 130.423C359.38 129.134 356.967 128.504 354.532 128.624C352.096 128.745 349.763 129.61 347.87 131.094C345.977 132.577 344.622 134.603 344 136.879L345.507 137.92C373.714 143.806 402.913 143.679 431.063 137.547C421.907 128.262 410.648 119.977 397.485 118.261Z" fill="#F0F0F0"/>
-<path d="M484.485 130.261C471.322 128.544 456.295 135.409 452.884 147.795C452.024 145.593 450.461 143.712 448.42 142.423C446.38 141.134 443.967 140.504 441.532 140.624C439.096 140.745 436.763 141.61 434.87 143.094C432.977 144.577 431.622 146.603 431 148.879L432.507 149.92C460.714 155.806 489.913 155.679 518.063 149.547C508.907 140.262 497.648 131.977 484.485 130.261Z" fill="#F0F0F0"/>
-<path d="M206.992 356.807C223.557 356.807 236.985 343.867 236.985 327.904C236.985 311.941 223.557 299 206.992 299C190.428 299 177 311.941 177 327.904C177 343.867 190.428 356.807 206.992 356.807Z" fill="#6C63FF"/>
-<path d="M206.721 380.174L206.965 327.745" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-<path d="M206.966 335.159C211.216 335.159 214.66 331.839 214.66 327.745C214.66 323.65 211.216 320.33 206.966 320.33C202.717 320.33 199.273 323.65 199.273 327.745C199.273 331.839 202.717 335.159 206.966 335.159Z" fill="#3F3D56"/>
-<path d="M206.815 360.058C206.815 360.058 205.823 337.272 183.279 339.822Z" fill="#3F3D56"/>
-<path d="M147.305 235.807C163.87 235.807 177.298 222.867 177.298 206.904C177.298 190.941 163.87 178 147.305 178C130.741 178 117.313 190.941 117.313 206.904C117.313 222.867 130.741 235.807 147.305 235.807Z" fill="#6C63FF"/>
-<path d="M147.324 381.892V207.01" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-<path d="M147.324 214.341C151.573 214.341 155.018 211.022 155.018 206.927C155.018 202.832 151.573 199.513 147.324 199.513C143.075 199.513 139.63 202.832 139.63 206.927C139.63 211.022 143.075 214.341 147.324 214.341Z" fill="#3F3D56"/>
-<path d="M123.694 340.87C127.943 340.87 131.387 337.55 131.387 333.455C131.387 329.361 127.943 326.041 123.694 326.041C119.445 326.041 116 329.361 116 333.455C116 337.55 119.445 340.87 123.694 340.87Z" fill="#3F3D56"/>
-<path d="M146.4 272.5C146.4 272.5 145.301 249.719 122.769 252.367Z" fill="#3F3D56"/>
-<path d="M147.324 309.382C147.324 309.382 146.225 286.601 123.694 289.249Z" fill="#3F3D56"/>
-<path d="M147.324 379.854C147.324 379.854 152.27 335.574 123.694 333.455" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-<path d="M171.427 271.32C175.676 271.32 179.12 268.001 179.12 263.906C179.12 259.811 175.676 256.492 171.427 256.492C167.178 256.492 163.733 259.811 163.733 263.906C163.733 268.001 167.178 271.32 171.427 271.32Z" fill="#3F3D56"/>
-<path d="M147.796 307.943C147.796 307.943 142.85 266.024 171.427 263.906" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-<path d="M48.6936 339.829C52.9427 339.829 56.3873 336.509 56.3873 332.414C56.3873 328.32 52.9427 325 48.6936 325C44.4446 325 41 328.32 41 332.414C41 336.509 44.4446 339.829 48.6936 339.829Z" fill="#3F3D56"/>
-<path d="M79.9924 301.807C96.5568 301.807 109.985 288.867 109.985 272.904C109.985 256.941 96.5568 244 79.9924 244C63.4281 244 50 256.941 50 272.904C50 288.867 63.4281 301.807 79.9924 301.807Z" fill="#6C63FF"/>
-<path d="M78.2941 379.609L80.0203 272.872" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-<path d="M80.0203 280.286C84.2694 280.286 87.7139 276.967 87.7139 272.872C87.7139 268.777 84.2694 265.457 80.0203 265.457C75.7712 265.457 72.3266 268.777 72.3266 272.872C72.3266 276.967 75.7712 280.286 80.0203 280.286Z" fill="#3F3D56"/>
-<path d="M79.532 303.063C79.532 303.063 78.8015 280.269 56.23 282.578Z" fill="#3F3D56"/>
-<path d="M78.3271 377.571C78.3271 377.571 83.9808 333.842 55.4421 331.295" stroke="#3F3D56" stroke-width="2" stroke-miterlimit="10"/>
-</g>
-<defs>
-<clipPath id="clip0_1_2">
-<rect width="550.71" height="567.986" fill="white"/>
-</clipPath>
-</defs>
-</svg>
-`);
+// function loadImages() {
+//     document.getElementsByClassName('developer-img')[0].insertAdjacentHTML('beforeend', DEVELOPER);
+//     document.getElementsByClassName('canvas-img')[0].insertAdjacentHTML('beforeend', CANVAS);
+// }
 
-// const links = Array.from(document.querySelectorAll('.navigation-links a'));
+view.loadImages(DEVELOPER, CANVAS);
 
-// links.forEach(link => {
-//     link.addEventListener('mouseover', () => {
-//         document.getElementById('right-eye').setAttribute('ry', 1);
-//         document.getElementById('right-brow').setAttribute('d', 'M 57 40 Q 61 38 64 40');
-//         document.getElementById('smile').setAttribute('d', 'M 38 62 C 43 67, 53 67, 62 59');
+// let roleElm = document.getElementsByClassName('role')[0];
+// let cursorElm = document.getElementsByClassName('cursor')[0];
+// let CHAR_DELAY = 400;
+
+// async function renderString(str) {
+//     cursorElm.classList.add('blink');
+//     await runAfterDelay(() => cursorElm.classList.remove('blink'), 2000);
+
+//     for (let i = 0; i < str.length; i++) {
+//         highlightKeys(str[i]);
+//         await renderChar(str[i]);
+//     }
+
+//     cursorElm.classList.add('blink');
+//     await runAfterDelay(() => cursorElm.classList.remove('blink'), 2000); 
+
+//     for (let i = 0; i < str.length; i++) {
+//         highlightBackspace();
+//         await clearLastChar();
+//     }
+// }
+
+// function renderChar(char) {
+//     return runAfterDelay(() => roleElm.textContent += char, CHAR_DELAY);
+// }
+
+// function clearLastChar() {
+//     return runAfterDelay(() => {
+//         roleElm.textContent = roleElm.textContent.substring(0, roleElm.textContent.length - 1);
+//     }, CHAR_DELAY);   
+// }
+
+// function runAfterDelay(callback, delay) {
+//     return new Promise((resolve) => {
+//         setTimeout(() => {
+//             callback();
+//             resolve();
+//         }, delay);
 //     });
+// }
 
-//     link.addEventListener('mouseout', () => {
-//         document.getElementById('right-eye').setAttribute('ry', 4);
-//         document.getElementById('right-brow').setAttribute('d', 'M 57 35 Q 61 33 64 35');
-//         document.getElementById('smile').setAttribute('d', 'M 38 62 C 43 67, 53 67, 62 62');
+// async function renderStrings(strings) {
+//     for (let i = 0; i < strings.length; i++) {
+//         await renderString(strings[i]);
+//     }
+//     await renderStrings(strings);
+// }
+
+
+// function highlightKeys(char) {
+//     let keys = getKeys(char);
+
+//     keys.forEach(key => {
+//         document.getElementById(key).style.fill = '#6C63FF';
+//         setTimeout(() => document.getElementById(key).style.fill = 'white', CHAR_DELAY);
+//         runAfterDelay(() => () => document.getElementById(key).style.fill = 'white', )
 //     });
-// });
+// }
 
-let roleElm = document.getElementsByClassName('role')[0];
-let cursorElm = document.getElementsByClassName('cursor')[0];
-let CHAR_DELAY = 400;
-
-async function renderString(str) {
-    cursorElm.classList.add('blink');
-    await new Promise((resolve, reject) => {
-        setTimeout(() => {
-            cursorElm.classList.remove('blink');
-            resolve();
-        }, 2000);
-    });
-
-    for (let i = 0; i < str.length; i++) {
-        highlightKeys(str[i]);
-        await renderChar(str[i]);
-    }
-
-    cursorElm.classList.add('blink');
-
-    await new Promise((resolve, reject) => {
-        setTimeout(() => {
-            cursorElm.classList.remove('blink');
-            resolve();
-        }, 2000);
-    });
-
-    for (let i = 0; i < str.length; i++) {
-        highlightBackspace();
-        await clearLastChar();
-    }
-    console.log('rendered', str);
-}
-
-function renderChar(char) {
-    return new Promise((resolve, reject) => {
-        setTimeout(() => {
-            roleElm.textContent += char;
-            resolve();
-        }, CHAR_DELAY);
-    })
-}
-
-function clearLastChar() {
-    return new Promise((resolve, reject) => {
-        setTimeout(() => {
-            roleElm.textContent = roleElm.textContent.substring(0, roleElm.textContent.length - 1);
-            resolve();
-        }, CHAR_DELAY);
-    })
-}
-
-async function renderStrings(strings) {
-    for (let i = 0; i < strings.length; i++) {
-        await renderString(strings[i]);
-    }
-    await renderStrings(strings);
-}
+// function highlightBackspace() {
+//     document.getElementById('Key_Backspace').style.fill = '#6C63FF';
+//     setTimeout(() => document.getElementById('Key_Backspace').style.fill = 'white', CHAR_DELAY);
+// }
 
 
-function highlightKeys(char) {
-    let keys = getKeys(char);
-
-    keys.forEach(key => {
-        document.getElementById(key).style.fill = '#6C63FF';
-        setTimeout(() => document.getElementById(key).style.fill = 'white', CHAR_DELAY);
-    });
-}
-
-function highlightBackspace() {
-    document.getElementById('Key_Backspace').style.fill = '#6C63FF';
-    setTimeout(() => document.getElementById('Key_Backspace').style.fill = 'white', CHAR_DELAY);
-}
-
-
-function getKeys(char) {
-    if (char === ' ')
-        return ['Key_Spacebar'];
-    else if (char.charCodeAt() >= 65 && char.charCodeAt() <= 90)
-        return [`Key_${char.toLowerCase()}`, 'Key_Shift'];
-    else if (char.charCodeAt() >= 97 && char.charCodeAt() <= 122)
-        return [`Key_${char}`];
-    return [];
-}
+// function getKeys(char) {
+//     if (char === ' ')
+//         return ['Key_Spacebar'];
+//     else if (char.charCodeAt() >= 65 && char.charCodeAt() <= 90)
+//         return [`Key_${char.toLowerCase()}`, 'Key_Shift'];
+//     else if (char.charCodeAt() >= 97 && char.charCodeAt() <= 122)
+//         return [`Key_${char}`];
+//     return [];
+// }
 
 let strings = ['Frontend Engineer', 'Artist', 'Designer'];
-renderStrings(strings);
+// renderStrings(strings);
+
+(async function renderStringsInLoop(strings) {
+    await view.renderStrings(strings);
+    renderStringsInLoop(strings);
+})(strings);
