+++ src/utils.js
@@ -1,0 +1,8 @@
+export const runAfterDelay = (callback, delay) => {
+    return new Promise((resolve) => {
+        setTimeout(() => {
+            callback();
+            resolve();
+        }, delay);
+    });
+}
